Cambios pendientes antes del pull

+++ bajastecnicas/bajas/views.py
@@ -1,4 +1,5 @@
 import datetime
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib.auth.decorators import login_required
 from django.db import models
@@ -7,6 +8,12 @@ from datetime import datetime
 from .models import Bajas
 from .forms import BajasForm
 from django.core.paginator import Paginator
+
+from xhtml2pdf import pisa
+from django.template.loader import render_to_string
+from django.contrib.sites.shortcuts import get_current_site
+import base64
+from django.core.files.storage import default_storage
 
 # Create your views here.
 def bajas_list(request):
@@ -79,7 +86,8 @@ def bajas_list(request):
         bajass = bajass.filter(anexo_a3=anexo_a3)
 
     #paginador 
-
+ 
+    bajass = bajass.order_by('id')
     paginator = Paginator(bajass, 3)
     page_number = request.GET.get('page')
     bajas = paginator.get_page(page_number)
@@ -105,11 +113,19 @@ def crear_baja(request):
 
     return render(request, 'bajas/crear_bajas.html', {'form': form})
 
-def eliminar_baja(request,id):
+def eliminar_baja(request, id):
+    is_ledesma = request.user.username.lower() == 'ledesma'
     baja = get_object_or_404(Bajas, pk=id)
-    if request.method == 'POST':
-        baja.delete()
-        return redirect('bajas:list')
+
+    if request.method == 'POST':  # verifica si el método es POST
+        if is_ledesma:  # Si el usuario es 'ledesma', permite la eliminación
+            baja.delete()
+            return redirect('bajas:list')  # Redirige a la lista después de eliminar
+        else:
+            return render(request, 'bajas/acceso_denegado.html')  # Si no es 'ledesma', muestra el acceso denegado
+
+    # Si no es POST, también puede manejarse aquí (aunque generalmente los formularios de eliminación usan POST)
+    return render(request, 'bajas/acceso_denegado.html') 
 
 def editar_baja(request, id):
     baja = get_object_or_404(Bajas, pk=id)
@@ -123,5 +139,29 @@ def editar_baja(request, id):
         form = BajasForm(instance=baja)
     return render(request, 'bajas/editar_bajas.html', {'form': form, 'baja': baja, 'is_ledesma': is_ledesma})
 
+def image_to_base64(image_path):
+    with default_storage.open(image_path, 'rb') as img_file:
+        return base64.b64encode(img_file.read()).decode('utf-8')
 
+def exportar_pdf(request, id):
+    baja = get_object_or_404(Bajas, id=id)
+    if baja.foto:
+        img_data = image_to_base64(baja.foto.path)
+        img_tag = f"data:image/jpeg;base64,{img_data}"
+    else:
+        img_tag = None
+
+    # Generar el HTML que va a ser convertido a PDF
+    context = {'baja': baja, 'img_tag': img_tag}
+    html_string = render_to_string('bajas/pdf_template.html', context)
+
+    response = HttpResponse(content_type='application/pdf')
+    response['Content-Disposition'] = f'attachment; filename="baja_{baja.id}.pdf"'
+
+    pisa_status = pisa.CreatePDF(html_string, dest=response)
+
+    if pisa_status.err:
+        return HttpResponse('Hubo un error al generar el PDF', status=400)
     
+    return response
+
